beginnings of redesign of JSLoader.dynamic_generate which ONLY loads on-demand from the absolute minimum OS file-walking.
this to ensure that a move to a distributed filesystem has minimum impact.

the current system (jumpscale.py) has a MASSIVE front-loading impact
of hitting the entire filesystem with loading absolutely every single
module using the standard python "import" system, which has the unfortunate
side-effect of hitting absolutely every single entry in sys.path.

in "develop" mode, where sys.path is patched, that means that there will
be MULTIPLE hits (not just the standard sys.path entries) of every plugin
with a search for that module.

the plan is to do a pre-walk of the filesystem and store the child
module import dependencies *in the parent python file that imports it*,
in an __jsdeps__ dictionary.  hand-creation of these entries is
being done (for now) in order to be able to get "from here to there".