add option to set the HOSTCFGDIR from the command-line
it is a bit of a hack but works.  otherwise the "-c" option has to be
laboriously added to absolutely every single command in apps/js_*

it really should be done as a library function... but unfortunately
the only sensible place to import from is... jumpscale... and unnfortunately
the only thing that is allowed to be imported from that is "j". oops.
and if you "import j", it is set up right there, right then, WITHOUT
ever having had time to get the config options parsed.

globals are a bitch.