issue #33 mitigation: tests run in new (blank) config each time
the unit tests (some of which are destructive) are now isolated to their
own subdirectory each time the unit tests are run.

unit tests that need special configuration options should spawn a new
process, in which the HOSTCFGDIR environment is pre-set, and the config
also pre-set-up.

global variable imports baaaaaad.