major flaw in bash ~/.profile_js handling: issue #70
environment variable was being deleted from MEMORY ONLY, it was not
actually being deleted from the actual ~/.profile_js file.

calling self.profileJS.save(True), which is also done when the profile
has a key ADDED, fixes the problem.

there may be other instances where this is an issue, in other profile
handling.