j.sal.processmanager.testSync error condition could never work
this unit test could never have succeeded (indicating a critical issue
with the code), as exceptions that could occur could never be properly
detected and passed to the caller.

eco.toJson had been changed to a property

the j.logger.log function no longer existed

the teardown (in the finally clause) could itself *cause* exceptions
(when an exception occurred in the caller).

the read from the pipe returns no data, causing the serialisation
to fail (throwing an exception in the finally block)

all a bit of a mess, really.  mostly sorted now: at least the test works