grrrr, a ton of classes will have been COMPLETELY missed
the JSLoader.py findJumpscaleLocationsInFile function is basically a load
of shit.  it should never have been written: standard python "compile"
or AST-walking should have been done, or even just importing the
module (using imputil) and doing object-walking looking for classes.

this bug - which has been there for a long time - was that if there
are multiple classes in a single file, locfound was not being reset
when the new class was found.  consequently, its __jslocation__ was
never found.

locfound is now reset when a new class is encountered, and it works.