if digitalme not installed, entire app breaks issue #94
the checking on j.application.schemas had not been tested for when
digitalme has not been installed.

if it does not exist then there is absolutely no way to even start up
any apps.  these modifications at least allow "silent fail", and
if the AlertHandler log is ever used it THEN throws a runtime exception