started investigating the SerializerFactory
initially wanted to just convert to j-less, it was however such a mess
that this commit has turned unfortunately into one big cleanup and
friday night investigation.

two issues have been raised (so far...) issue #97 where it was found
that the yaml serializer test appears to be failing (which needs full
investigation... just not over the weekend...) and issue 98 where
the Serializer class is commented out... so the get function is
non-functional (and no comments or documentation indicating that
were present).  so....