issue #106 - configmanager.keyname was not initialised
in configmanager in the path property, if the path is blank after
obtaining it from the dictionary, self.init is called as that is a
clear sign that initialisation has not been carried out.

however... the keyname can also be blank, and there was no test for
whether this was the case or not.  in j.data.nacl.test() by calling
NACL.get *BEFORE* configmanager has been initialised, a BLANK
sshkeyname was being passed in.

checking if it is blank and if so calling initialisation, although
annoying as it is interactive, at least gets the keyname set up.

at least two things to review here